Homogenizing Counter Contracts
The current RUST contract had a private method which simply warned
about the possibility of a i8 under/overflow. I have changed it
for a method that checks if the value is about to under/overflow.
This still represents well what a private method is, while presenting
the users with a better use case or private methods.